+++ src/controllers/productController.js
@@ -1,106 +1,18 @@
-// const { PrismaClient } = require("@prisma/client");
-// const uploadService = require("../services/uploadService");
-
-// const prisma = new PrismaClient();
-// const productController = {};
-
-// productController.createProduct = async (req, res, next) => {
-//   const { productImage, productType, productName, productDetail, price } =
-//     req.body;
-
-//   try {
-//     const newProduct = await prisma.product.create({
-//       data: {
-//         productImage,
-//         productType,
-//         productName,
-//         productDetail,
-//         price: parseFloat(price),
-//       },
-//     });
-
-//     res.json(newProduct);
-//   } catch (error) {
-//     res.status(500).json({ error: "Could not create product" });
-//   }
-// };
-
-// productController.getAllProduct = async (req, res, next) => {
-//   try {
-//     const products = await prisma.product.findMany();
-//     res.json(products);
-//   } catch (err) {
-//     console.log(err);
-//     res.status(500).json({ message: err.message });
-//   }
-// };
-
-// productController.getProductById = async (req, res, next) => {
-//   const { productId } = req.params;
-
-//   try {
-//     const product = await prisma.product.findUnique({
-//       where: {
-//         id: parseInt(productId),
-//       },
-//     });
-//     if (!product) {
-//       return res.status(404).json({ err: "Product not found" });
-//     }
-//     res.status(200).json(product);
-//   } catch (err) {
-//     console.log(err);
-//     res.status(500).json({ message: "no fetch product" });
-//   }
-// };
-
-// productController.updateProduct = async (req, res, next) => {
-//   const { id } = req.params;
-//   const { productImage, productType, productName, productDetail, price } =
-//     req.body;
-
-//   try {
-//     const updatedProduct = await prisma.product.update({
-//       where: { id: parseInt(id) },
-//       data: {
-//         productImage,
-//         productType,
-//         productName,
-//         productDetail,
-//         price: parseFloat(price),
-//       },
-//     });
-
-//     res.json(updatedProduct);
-//   } catch (error) {
-//     res.status(500).json({ error: "Could not update product" });
-//   }
-// };
-
-// productController.deleteProduct = async (req, res, next) => {
-//   const { id } = req.params;
-
-//   try {
-//     const deletedProduct = await prisma.product.delete({
-//       where: { id: parseInt(id) },
-//     });
-
-//     res.json(deletedProduct);
-//   } catch (error) {
-//     res.status(500).json({ error: "Could not delete product" });
-//   }
-// };
-
-// module.exports = productController;
-
 const { PrismaClient } = require("@prisma/client");
 const uploadService = require("../services/uploadService");
 
 const prisma = new PrismaClient();
 const productController = {};
 
+// เพิ่มcheck admin
 productController.createProduct = async (req, res, next) => {
   try {
+    if (!req.user.isAdmin) {
+      return res
+        .status(403)
+        .json({ error: "Access denied. User is not admin." });
+    }
+
     const { productType, productName, productDetail, productPrice } = req.body;
 
     const result = await uploadService.upload(req.file.path);
@@ -151,7 +63,11 @@ productController.getProductById = async (req, res, next) => {
 
 productController.updateProduct = async (req, res, next) => {
   const { productId } = req.params;
-  // const { productName, productType, productDetail, productPrice } = req.body;
+
+  if (!req.user.isAdmin) {
+    return res.status(403).json({ error: "Access denied. User is not admin." });
+  }
+
   req.body.price = +req.body.productPrice;
   delete req.body.productPrice;
 
@@ -177,6 +93,10 @@ productController.updateProduct = async (req, res, next) => {
 
 productController.deleteProduct = async (req, res, next) => {
   const { productId } = req.params;
+
+  if (!req.user.isAdmin) {
+    return res.status(403).json({ error: "Access denied. User is not admin." });
+  }
 
   try {
     await prisma.product.delete({
